+++ services/horizon/internal/db2/history/mock_transactions_batch_insert_builder.go
@@ -1,15 +1,16 @@
 package history
 
 import (
-	"github.com/stellar/go/ingest/io"
 	"github.com/stretchr/testify/mock"
+
+	"github.com/stellar/go/ingest"
 )
 
 type MockTransactionsBatchInsertBuilder struct {
 	mock.Mock
 }
 
-func (m *MockTransactionsBatchInsertBuilder) Add(transaction io.LedgerTransaction, sequence uint32) error {
+func (m *MockTransactionsBatchInsertBuilder) Add(transaction ingest.LedgerTransaction, sequence uint32) error {
 	a := m.Called(transaction, sequence)
 	return a.Error(0)
 }