+++ services/horizon/internal/ingest/history_archive_adapter_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 
 	"github.com/stellar/go/historyarchive"
-	"github.com/stellar/go/ingest/io"
+	"github.com/stellar/go/ingest"
 	"github.com/stellar/go/xdr"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
@@ -27,9 +27,9 @@ func (m *mockHistoryArchiveAdapter) BucketListHash(sequence uint32) (xdr.Hash, e
 	return args.Get(0).(xdr.Hash), args.Error(1)
 }
 
-func (m *mockHistoryArchiveAdapter) GetState(ctx context.Context, sequence uint32) (io.ChangeReader, error) {
+func (m *mockHistoryArchiveAdapter) GetState(ctx context.Context, sequence uint32) (ingest.ChangeReader, error) {
 	args := m.Called(ctx, sequence)
-	return args.Get(0).(io.ChangeReader), args.Error(1)
+	return args.Get(0).(ingest.ChangeReader), args.Error(1)
 }
 
 func TestGetState_Read(t *testing.T) {