+++ services/horizon/internal/ingest/processor_runner_test.go
@@ -2,18 +2,20 @@ package ingest
 
 import (
 	"context"
+	"io"
 	"reflect"
 	"testing"
 
 	"github.com/guregu/null"
-	"github.com/stellar/go/ingest/io"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/mock"
+
+	"github.com/stellar/go/ingest"
 	"github.com/stellar/go/ingest/ledgerbackend"
 	"github.com/stellar/go/network"
 	"github.com/stellar/go/services/horizon/internal/db2/history"
 	"github.com/stellar/go/services/horizon/internal/ingest/processors"
 	"github.com/stellar/go/xdr"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/mock"
 )
 
 func TestProcessorRunnerRunHistoryArchiveIngestionGenesis(t *testing.T) {
@@ -98,6 +100,11 @@ func TestProcessorRunnerRunHistoryArchiveIngestionHistoryArchive(t *testing.T) {
 	historyAdapter.On("BucketListHash", uint32(63)).
 		Return(xdr.Hash([32]byte{0, 1, 2}), nil).Once()
 
+	m := &ingest.MockChangeReader{}
+	m.On("Read").Return(ingest.GenesisChange(network.PublicNetworkPassphrase), nil).Once()
+	m.On("Read").Return(ingest.Change{}, io.EOF).Once()
+	m.On("Close").Return(nil).Once()
+
 	historyAdapter.
 		On(
 			"GetState",
@@ -105,9 +112,7 @@ func TestProcessorRunnerRunHistoryArchiveIngestionHistoryArchive(t *testing.T) {
 			uint32(63),
 		).
 		Return(
-			&io.GenesisLedgerStateReader{
-				NetworkPassphrase: network.PublicNetworkPassphrase,
-			},
+			m,
 			nil,
 		).Once()
 
@@ -146,7 +151,7 @@ func TestProcessorRunnerRunHistoryArchiveIngestionHistoryArchive(t *testing.T) {
 		Return(mockClaimableBalancesBatchInsertBuilder).Once()
 
 	q.MockQAccounts.On("UpsertAccounts", []xdr.LedgerEntry{
-		xdr.LedgerEntry{
+		{
 			LastModifiedLedgerSeq: 1,
 			Data: xdr.LedgerEntryData{
 				Type: xdr.LedgerEntryTypeAccount,
@@ -269,7 +274,7 @@ func TestProcessorRunnerBuildChangeProcessor(t *testing.T) {
 		historyQ: q,
 	}
 
-	stats := &io.StatsChangeProcessor{}
+	stats := &ingest.StatsChangeProcessor{}
 	processor := runner.buildChangeProcessor(stats, ledgerSource, 123)
 	assert.IsType(t, &groupChangeProcessors{}, processor)
 
@@ -321,7 +326,7 @@ func TestProcessorRunnerBuildTransactionProcessor(t *testing.T) {
 		historyQ: q,
 	}
 
-	stats := &io.StatsLedgerTransactionProcessor{}
+	stats := &processors.StatsLedgerTransactionProcessor{}
 	ledger := xdr.LedgerHeaderHistoryEntry{}
 	processor := runner.buildTransactionProcessor(stats, ledger)
 	assert.IsType(t, &groupTransactionProcessors{}, processor)